+++ packages/flutter_adaptive_cards_fs/test/utils/test_utils.dart
@@ -302,6 +302,7 @@ Widget getTestWidgetFromMap({
     // so dislable them in the golden images
     showDebugJson: false,
     initData: initData,
+    onChange: onChange,
     hostConfigs: hostConfigs ?? HostConfigs(),
     listView: listView,
   );